bug fixed for spblib error

+++ Metis/SpaceGroup/GenerateSmallerUnitCell.py
@@ -2,12 +2,9 @@
 from Metis.Base.TspaceToolbox import TspaceToolbox
 from Metis.Espresso.QE2Spg import QE2Spg
 from Metis.SpaceGroup.SpaceGroup import SpaceGroup
-#from Metis.SpaceGroup.GenerateWyckoffPositionsList \
-#        import GenerateWyckoffPositionsList
 from Metis.SpaceGroup.IdentifySubIndex \
         import IdentifySubIndex
 import os
-from fractions import Fraction
 
 
 class GenerateSmallerUnitCell(TspaceToolbox):
@@ -24,12 +21,20 @@ class GenerateSmallerUnitCell(TspaceToolbox):
 
     def set_atom_data(self):
         #
-        # coordinate is not conventional unit but primitive unit
+        # in this case, smaller unit cell is not create case,
+        #   because spglib cannot identify space group
         #
-        lattice_length = self.atom_info['lattice_length']
-        lattice_angle = self.atom_info['lattice_angle']
-        self.lattice_type = self.atom_info['lattice_type']
-        atomic_positions = self.atom_info['atomic_positions']
+        if self.atom_info is None:
+            self.identified_spg = False
+            return
+        else:
+            #
+            # coordinate is not conventional unit but primitive unit
+            #
+            lattice_length = self.atom_info['lattice_length']
+            lattice_angle = self.atom_info['lattice_angle']
+            self.lattice_type = self.atom_info['lattice_type']
+            atomic_positions = self.atom_info['atomic_positions']
         self.generate_prim_crystal_in(lattice_length,
                                       lattice_angle,
                                       atomic_positions)
@@ -80,12 +85,20 @@ class GenerateSmallerUnitCell(TspaceToolbox):
         #
         # transform primitive cell to conventional cell
         #
-        self.spg = SpaceGroup(filename).get_conventional_cell()
+        space_group = SpaceGroup(filename)
+        self.identified_spg = True
+        if space_group.spg is None:
+            self.identified_spg = False
+            return
+        self.spg = space_group
+        ispg = self.spg.ispg
+        self.spg.get_conventional_cell()
+
         if os.path.isfile(filename):
             os.remove(filename)
-        
+
         #
-        #  identified reduced cell 
+        #  identified reduced cell
         #
 
         #
@@ -105,9 +118,24 @@ class GenerateSmallerUnitCell(TspaceToolbox):
             target_list = atom_info['wyckoff_letter']
             element = atom_info['element']
             sub_index = IdentifySubIndex(natoms=natoms,
-                                        ispg=self.spg.ispg,
-                                        target_list=target_list).sub_index
+                                         ispg=ispg,
+                                         target_list=target_list).sub_index
+            if sub_index is None:
+                print('===== sub_index Error(GenerateSmallerUnitCell) =====')
+                print(' filename = {}'.format(self.filename))
+                print(' identified spg ...', end='')
+                if self.identified_spg:
+                    print('ok')
+                else:
+                    print('ng')
+                print(' natoms = {}'.format(natoms))
+                print(' ispg = {}'.format(self.spg.ispg))
+                print(' target_list = ', end='')
+                print(target_list)
+                exit()
             self.ispg = self.spg.ispg
             self.sub_index = sub_index+1
-            self.dirname = '{0}{1}_{2}_{3}'.format(element, natoms, self.spg.ispg, sub_index+1)
+            self.dirname = '{0}{1}_{2}_{3}'.format(element, natoms,
+                                                   self.spg.ispg,
+                                                   sub_index+1)
             self.compound_name = '{0}{1}'.format(element, natoms)
